fix segfaults and improve

+++ lora-esp32/src/proc/lTRX.h
@@ -1,2 +1,2 @@
-void declTRX(int msgType, String data, int pktRS, int pktSNR);
+void declTRX(int msgType, String data, int pktRS, float pktSNR);
 void lTRXctl();
